+++ src/main/java/wooteco/subway/exception/constant/BlankArgumentException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.constant;
 
 public class BlankArgumentException extends IllegalArgumentException {
 }
