+++ src/main/java/wooteco/subway/exception/constant/NotExistException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.constant;
 
 public class NotExistException extends IllegalArgumentException{
 }
