+++ src/test/java/wooteco/subway/domain/StationTest.java
@@ -3,7 +3,7 @@ package wooteco.subway.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import wooteco.subway.exception.BlankArgumentException;
+import wooteco.subway.exception.constant.BlankArgumentException;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
